test-requirements.txt
django-debug-toolbar
selenium==4.*
coverage>=7.8,<8
